Bug fixes
Fixed bug where it crashes when the server socket is closed. Now, if it's not connected it attempts to connect once every time it's supposed to send a command. Also, fixed update_location to actually use Nodes from the pathfinder node list. Was creating new nodes with the right coordinates but unconnected to each other, so it failed to find a path every time.